Move versioning contract
